update: 좌표 필드값의 자료형을 double로 수정

+++ app/src/main/java/com/example/travelhelper/data/model/DestinationDetail.kt
@@ -4,6 +4,6 @@ data class DestinationDetail(
     val name: String,
     val address: String,
     val tel: String,
-    val x: Float,
-    val y: Float
+    val x: Double,
+    val y: Double
 )
